requirements.txt
blinker==1.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.5
phonenumbers==8.10.23
py3dns==3.2.1
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
twilio==6.33.1
urllib3==1.25.7
validate-email==1.3
Werkzeug==0.16.0
